[10]  Refactor lexer to extract fsm logic
closes #10
